R4R: CheckTx AnteHandler for Ethereum Tx Messages (#505)
* Rename Ethereum tx message
* Use new tx decoder in the ethermint app
* Update ante handler to prevent spam/dos
* Update ethereum msg signing/verification logic
* Implement secp256k1 key types
* Remove pointer from To method
* Move sig check to after inartistic gas check
* Add comment on chainID parsing
* Updated validateIntrinsicGas godoc
* Implement Fee method on eth tx msg
* Add reference to spec for recoverEthSig
* Upgrade TM to v0.27.0

+++ x/evm/types/state_object.go
@@ -176,10 +176,10 @@ func (so *stateObject) SetNonce(nonce uint64) {
 		prev:    so.account.Sequence,
 	})
 
-	so.setNonce(int64(nonce))
+	so.setNonce(nonce)
 }
 
-func (so *stateObject) setNonce(nonce int64) {
+func (so *stateObject) setNonce(nonce uint64) {
 	so.account.Sequence = nonce
 }
 
@@ -249,7 +249,7 @@ func (so *stateObject) CodeHash() []byte {
 
 // Nonce returns the state object's current nonce (sequence number).
 func (so *stateObject) Nonce() uint64 {
-	return uint64(so.account.Sequence)
+	return so.account.Sequence
 }
 
 // Code returns the contract code associated with this object, if any.
